Simplify the view and presenter - Site location
For reference this change includes some changes to the factories I needed to make to support the new tests.

I essentially copied and pasted my changes from what @cintamani has done in the WEX engine, and at some point it would benefit updating all the factories to match the refactoring done in the WEX engine.