Simplify CSS
The existing CSS was copied wholesale from WEX. What Ed was not aware of when he copied it is the context for why it was added.

Viewing the confirmation letter was originally directly tied with the bulk export confirmation letter print. The original design for the export was to view all the letters to be printed, and the user would then hit print in the browser.

Hence there was a lot of html designed to handle viewing numerous letters at once. To get it all to work they followed some posts which detailed how to effectively reset the HTML, and then apply your changes on top.

So to simplify things I have cut out everything that is not related to what we do now. With this updated CSS the letter prints in the same way as the current letter apart from the UL's (in the current letter the bullet is aligned to the left of the content which is clearly just wrong!)