Remove presenter's private t() helper
With the removal of `web_page_title()` helper from the presenter there are only a couple of places where the `t()` helper method was being used.

Also on review I have found it more confusing to remember that calls to `t()` in the presenter are different to calls to `t()` in a view.

So this change is my decision to remove the helper and just use `I18n.t()` directly where its needed in the presenter.