Fix issues found with confirmation letter (#135)
When the change for the confirmation letter generation went through QA a number of issues were found.

- Date format is different to the original
- Logo missing from letter
- Filename is different to the original
- Exemptions information is missing code and shows status rather than expiry date
- translation missing error for sole trader business type

The code was also pretty much taken as is from the previous waste exemptions project but we believe there were some refactorings and changes that could be made to make it more consistent with our current approach and standards.

We also needed to remove the wkhtmltopdf-binary dependency added in the original change. This is because we haven't included this additional dependency when generating pdf's previously. It would be worth looking at in the future as it means we control our dependency on wkhtmltppdf rather than rely on web-ops. However we couldn't use this particular one because it has no license (😡).

So to bring this in line with all our other projects, this change

- removes the wkhtmltopdf-binary gem
- adds the bash script used to start both xvfb and wkhtmltopdf (both are needed in our linux environments)
- updates the wickedpdf initializer to set wickedpdf's exe path as the path to the script