Remove wickedpdf gem and configs
part of: https://eaflood.atlassian.net/browse/RUBY-223

The wicked_pdf gem, initializer config and script are already part of the engine code and dependency of this application.
Requiring and installing the gem twice was causing an error generation in the PDF recorded in the ticket above, for which a PDF was generated with wrong content.