Update to include site address area (#194)
https://eaflood.atlassian.net/browse/RUBY-25

The existing service has the ability to lookup the EA admin area for a site using the easting and northing values. This is currently out of action whilst we await a fix (the external service we call changed supplier and we have been awaiting support for what the new query should be).

However it has been spotted in our work on migrating the old data to the new schema that we don't have a field for this. So on the basis we want to retain this data, and that we expect to implement the lookup feature in the new service once the issue is resolved, this change adds update the WEX engine to bring in a version that adds `area` to the **address** and **transient_address** models.

It also updates the registration details view to include this information. (We display the area next to the site address using the title **Area** to remain consistent with how the area is currently shown in the existing system)