Rename the reports table and add data dates info
From https://eaflood.atlassian.net/browse/RUBY-392

This renames the defra_ruby_exporters_bulk_export_files table to reports_generated_reports.
This also add more information to the reports, which are the dates of which the data refer to.
I find it a useful info to have there, and I have updated the code that deal with those dates to reference those fields so that we remove the needs of parsing a filename that might or might not contains these info.