Testing travis build
The travis build seems to be failing for the back office. Error

```bash
0.00s$ cp config/database.travis.yml config/database.yml
before_script.6
2.95s$ RAILS_ENV=test bundle exec rake db:create --trace
The HashDiff constant used by this gem conflicts with another gem of a similar name.  As of version 1.0 the HashDiff constant will be completely removed and replaced by Hashdiff.  For more information see https://github.com/liufengyun/hashdiff/issues/45.
** Invoke db:create (first_time)
** Invoke db:load_config (first_time)
** Execute db:load_config
** Execute db:create
could not connect to server: No such file or directory
	Is the server running locally and accepting
	connections on Unix domain socket "/var/run/postgresql/.s.PGSQL.5432"?
```

This is an investigative branch and PR to try and ascertain what is wrong.