Downgrarde Postgres version to 9.6
Following Trravis build failures we have noticed that the travis VM have updated the Postgres version from 9.6 to 10+ when the errrors started appearing.
This downgrade fixes the situation for now.
I am not 100% sure about the real issue, since the current PG gem version is suppose to support postgres 10+.

A ticket have been created to follow up on this:
https://github.com/DEFRA/ruby-services-team/issues/30