Change return field for assisted mode
I was under the impression that since the `assistance_mode` was a string type field with no obvious constraint it would contain just the same string as the CSV. Instead seems like we populate the field to `full` when the registration is a fully assisted one while we leave it blank otherwise.
Fix the code so that we will return the string "fully assisted" when the field is setted to "full" and "unassisted" otherwise.