Implement EPR exports with new DefraRubyAws gem
From https://eaflood.atlassian.net/browse/RUBY-392

In order to remove the exporter gem from the code base we have decided to also implement the EPR report as part of the 392 ticket.

This code hence includes a new rake task for epr reports in the back-office-ta projects to generate a CSV containing all currently active exemptions and upload it to an AWS bucket.