Only list expireable exemptions
This excludes exemptions which are already expired or have been ceased or revoked.

Had to change how the factory works as assigning `exemptions` and `registration_exemptions` separately was resulting in different results when `registration.exemptions` and `registration.registration_exemptions` were called. Modifying the `registration_exemptions` which are created by calling `exemptions` earlier results in a more accurate result.