Build first reminder email (#229)
https://eaflood.atlassian.net/browse/RUBY-252

This is the first reminder email that digital users will get. It should be sent 4 weeks before the registration is due to expire.

This story covers building the mailer and putting the HTML and text into the code.

We also added a temporary rake task to send a reminder email. This is solely for test purposes while we are still in the process of building this feature and should be removed later.

Had to change how the factory works as assigning `exemptions` and `registration_exemptions` separately was resulting in different results when `registration.exemptions` and `registration.registration_exemptions` were called. Modifying the `registration_exemptions` which are created by calling `exemptions` earlier results in a more accurate result.

Using the `.located_by_grid_reference?` method meant updating the address factory to more accurately resemble an address which was located by grid reference. This resulted in one of the exporter tests breaking so had to fix that too.