Update bulk exporter schedule to be daily (#244)
Now we have completed the migrations work, we have been able to run the bulk export in pre-production with production volumes. It currently takes 23 minutes, and we believe that is perfectly acceptable as a daily rather than weekly job.

Hence this change which removes the frequency element of the bulk export schedule, setting it to run daily at 20:05.