Fix error when in confirmation letter for expired
https://eaflood.atlassian.net/browse/RUBY-453

Prior to this change if the registration contained an exemption with a state of `expired` it would error when the user attempted to view the confirmation letter.

A quick check of the code identified that the logic in the `ConfirmationLetterPresenter.registration_exemption_status()` method would attempt to generate a display version of the deregistered date for any exemption without a status of `active`.

The problem is that an expired exemption does not have a deregistered date hence the code was failing. So a quick tweak of the logic and 🎉 we have a working confirmation letter.