Add rake task to set the status of expired registration exemptions
From https://eaflood.atlassian.net/browse/RUBY-413

This adds a rake task and a service object to deal with setting the status of a registration exemption to expired when it is in an active status and the expire date is in the past.
After talking about expiring on the day of expire or the day before, we have decided that the registration is considered expired after the date have passed, hence this include a fix on our logic for "not_expired" registrations too.
One more thing part of this PR is the definition of the RegistrationExemption table name which allow us to use the AASM default defined scopes.