Break long emails onto multiple lines instead of overflowing
https://eaflood.atlassian.net/browse/RUBY-385

On the registration details page, if you have very long email addresses or are zoomed in with certain browsers, the emails can overlap and become unreadable.

This changes adds some CSS to break the emails onto multiple lines when necessary. The link still works and the text can be copied and pasted as a single line.