Update privacy policy (#270)
For whatever reason, the version of the privacy policy in the back office has not been updated inline with the version in the front office. So prior to this change it was still the old version prior to our GDPR review.

It's possibly the fact the link to the privacy policy in the footer does not exist in the back office, and hence it was overlooked. However with the new verbal privacy policy we do have a link, but without this change users will see the old content.

This change updates it to bring it inline with the front office, with GDPR reviewed content.