Prevent exports to fail if a required address is missing
From: https://eaflood.atlassian.net/browse/RUBY-523

We are currently facing an issue in regard of completed registrations not having all required addresses.
To prevent the Bulk and EPR exports to fail during our investigations, the code in this PR is a good candidate for an hot-fix release.
In the long term, we do prefer to be notified about the issue by someone flagging the info is missing from the export, rather than from a server error and hence failure to generate reports, hence this code is a good candidate to be merged and kept in the master code.