Move models to reports namespace
With moving the service to the `reports` namespace, it makes sense to also do the same for the models. However there are a number of them, so to try and make it clear that these classes relate to a single feature, this change moves them to `Reports::Boxi` rather than just into `Reports`.