Resolve issue with missing batch size env var
I had an issue where tests were passing locally but not on Travis. All the fails were in the boxi model serializers.

After some trial an error I tracked it down to the env var EXPORT_SERVICE_BATCH_SIZE being missing in the Travis config. Locally I had it set so everything was passing. In travis it wasn't so we got fails.

The reason is that the BOXI base serializer assumed that the env var existed, and it did in the old travis config before we renamed it. Our existing reports `base_serializer` had an additional check on the env var it used so that if not set it defaulted the value.

Rather than duplicate the checks, I've instead gone with a solution we normally use which is default it in `application.rb`. This means we don't need additional logic in the classes that use the value to check its been set.