Update renewal email to use magic link
https://eaflood.atlassian.net/browse/RUBY-492

Now that we have created the magic link route in https://github.com/DEFRA/waste-exemptions-engine/pull/245, we need to actually add it to the emails.

This will also require a feature flag so we don't deploy this version of the email until the whole renewals process is ready to release.