Move BOXI export to reports namespace (#287)
Spotted that the BOXI export was in its own namespace, and using its on env vars for config. On reflection as

- its an automated job expected to be run nightly
- it exports a bunch of data as CSV files and then them to AWS S3

It was felt it could be moved into the reports namespace to make things consistent. It does compress the csv files first before loading, but that is the only real difference.

So this change updates the code to move the boxi export logic to the `reports` namespace, and updates the naming of it and its env vars to be consistent with the other export functions.

It also resolves an issue where the BOXI export could fail if EXPORT_SERVICE_BATCH_SIZE was not set.

The reason is that the BOXI base serializer assumed that the env var existed, and it did in the old travis config before we renamed it. Our existing reports `base_serializer` had an additional check on the env var it used so that if not set it defaulted the value.

Rather than duplicate the checks, we instead went with a solution we normally use which is default it in `application.rb`. This means we don't need additional logic in the classes that use the value to check its been set.