Update renewal email to use magic link (#290)
https://eaflood.atlassian.net/browse/RUBY-492

Now that we have created the magic link route in https://github.com/DEFRA/waste-exemptions-engine/pull/245, we need to actually add it to the emails.

If a renew_token doesn't exist yet, we generate a new one first, and then get the link for the email.

This will also require a feature flag so we don't deploy this version of the email until the whole renewals process is ready to release. The feature toggle is currently disabled. We've discussed this and agreed that we should be able to override these with an env var, but the value in the feature toggle file should be the default.

This also includes a rake task for testing. So running `bundle exec rake reminder:magic_link` will send a magic link email for the newest Registration in the database.