Enable BOXI export job
https://eaflood.atlassian.net/browse/RUBY-460

Even though the BOXI universe is not ready, our export of the data to CSV and then AWS S3 is. It has been decided to have the export running ahead of the other changes, so we can be sure we are ready and working for when they are made.

Hence this change enables the BOXI export job.