Fix error in email_image_tag
The logo is being added to the back office, which is local for the mailer. Prior to this change the only logo we had was pulled from the engine.

We are using the `email_image_tag()` to add the image to the email body because it has stuff built in to automatically attach it to the email as well.

However on first use it was erroring, even though its supposed to handle both local images and those from the engine. Using `binding.pry` we tracked the error down.

```bash
[1] pry(#<#<Class:0x000055c9347abc20>>)> path
=> "/app/assets/images/logo_ea_color.png"
[2] pry(#<#<Class:0x000055c9347abc20>>)> Rails.root.join(path)
=> #<Pathname:/app/assets/images/logo_ea_color.png>
[3] pry(#<#<Class:0x000055c9347abc20>>)> Rails.root
=> #<Pathname:/vagrant/waste-exemptions-back-office-ta>
[4] pry(#<#<Class:0x000055c9347abc20>>)> File.join(Rails.root, path)
=> "/vagrant/waste-exemptions-back-office-ta/app/assets/images/logo_ea_color.png"
[5] pry(#<#<Class:0x000055c9347abc20>>)> File.exist?(File.join(Rails.root, path))
=> true
[6] pry(#<#<Class:0x000055c9347abc20>>)> File.exist?(Rails.root.join(path))
=> false
```

The problem is our assumption that `Rails.root.join(path)` would result in `"/vagrant/waste-exemptions-back-office-ta/app/assets/images/logo_ea_color.png". However it drops the root and just returns the original value of `path`.

Switching the call with `File.join(Rails.root, path)` resolves the problem. It can now find the file because its getting the full path, rather than trying to search in the engine.