Add ability to custom-parse attribute on Boxi serializer
From: https://eaflood.atlassian.net/browse/RUBY-556

We had the need to strip out newlines on multiline params such as `description` in the `addresses` table.
To do so, I have used some methaprogramming to create parse methods for an attribute so that we can easely twick those value in the future.