Add tests for second reminder mail
Whilst going through the code for this refactor, spotted that we had test coverage for the first email reminder, and at the time for the magic link reminder. But we didn't have coverage for the second.

This change just makes things consistent by adding some tests for second email reminder mailer.