Refactor email reminders to remove non-magic link (#334)
The original plan was to ship a version of the first reminder email that only asked users to re-register. This was because it was thought shipping renewals would take some time.

Thanks to the success of the team, we actually have renewals including the version of the email which includes a **magic link** ready to go. This means the old re-register email will never see the light of day.

However we still have code that makes distinctions between the re-register and magic link versions of the first email reminder. So this change is some housekeeping to clean up the logic around the first email reminder to remove those distinctions plus

- quick fix to the description of a unit test in the second reminder
- make the reminder feature toggles consistent
- add some missed tests for the second reminder email mailer