Don't require active status to display renewal link
We want to display the renew link when the registration is still in the renewal wndow. But we've spotted that the check to display the links requires the registration to still be active. However, some expired registrations could still be in the renewal window. In this case, the link would no longer display even when it should.

This PR fixes this issue.