Do not open renewals in new tab
We missed it on code review, but was found later in QA. The renewal workflow for an NCCC user should follow the same as resuming a new registration, or editing an existing one. It should happen in the existing tab and not open a new one.

Like resuming and editing, we continue to show the main menu in the header. The view confirmation feature opens the confirmation in PDF which does take the user out of the system. Hence we open that in a new tab, but we don't need to open resume, edit or renew in a new tab.