Make the registration responsible to know if it is renewable
This is a little refactoring on the action links helpers for renewal links.
First, we drop the check on the class name as the `can?` permission check will just fail unless the object is of the right class.
Second, we give responsibility to check if a registration is renewable to the registration itself.