Generate assisted digital renewal letter
https://eaflood.atlassian.net/browse/RUBY-578

For users who didn't register with an email address, we want to notify them by post.

This means we'll need to generate a PDF of the letter as part of the back office.

This PR implements the PDF generation and adds a way to test it (pending adding the proper ways to access it).