Update project following name change
We renamed the project in GitHub from **waste-exemptions-back-office-ta** to **waste-exemptions-back-office**. The **TA** was there just whilst we were working on updating the code as there already existed a `back-office` repo from the previous version of the service.

We've now archived and removed the old repo from GitHub, and completed our technical alignment work so felt the time was ready to rename this one and drop the `-ta`.

Most other integrations either updated automatically or could be done without changes to the code. However CodeClimate & Harkiri needed a few things updating.

Other than that this change is a case of searching the repo for `-ta` and removing it where found.