Clean old AD renewal letters files (#400)
* Clean old AD renewal letters files

From: https://eaflood.atlassian.net/browse/RUBY-635

This adds a new `deleted` status to the AD renewal letters export records.
The enum-defined `deleted!` method is then overriden to also delete the AWS file from the bucket before updating the status of the records to `deleted`. The dashboards result have been filtered out to not show records in a `deleted` status.