Rename and document service runner
IMHO I felt that the `ServiceOnScopeRunner` was more about running a thing for a specific period of time more than combining a scope and a service.

So I took the liberty of renaming it. Also as a sort of 'helper' class, I felt it could benefit from some documentation so those coming across it can understand it's intent.