Remove addressbase_url from application.rb
Didn't manage to spot this as part of the env vars housekeeping in 7b1bae2, but we don't need to capture as a rails config as we don't use it directly. Plus the waste exemptions engine initializer handles grabbing either the env var or the default value and passing it to the engine.