Remove dependence on env vars for some tests
During the housekeeping of the env vars in PR #427 it was noted that we default most env vars, but because our `schedule_spec.rb` tests specifically for the values of certain env vars we are still forced to set them across all environments.

We've decided that we can be confident calling `ENV["MY_VAR"]` will return the value of `MY_VAR` if set as an environment variable, and therefore the various cron jobs `schedule.rb` manages will have the correct time set.

So this change

- updates the `schedule_spec.rb` tests to just confirm the default values
- removes env vars that are defaulted in the app from `.travis.yml`