Add basic tests for other specs
Though the previous fix resolved the issue of recording the test coverage for the `lookups.rake` task accurately, it appears to have awoken the simeplcov leviathan and now its reporting now test coverage on the rest of our rake tasks.

So in all honesty, at this time this is just about resolving issues with test coverage, but as a bonus we at least know in our unit tests that these should run without raising any errors.