Move task_name definition to shared context
Originally thought it was not possible to specify the `:task_name` in the shared rake context, so defined it within the tests.

However a nudge from a colleague and I then worked out it was just an ordering issue. As long as `:task_name` is defined before `:subject` we get no error.