First step to resolve rake tests breaking others
Now we are finding that the rake tests seem to be affecting others in cases where data is created. Any records created in the rake tests don't appear to be cleaned properly which means they remain and effect the results of others.

Though it talks about using DatabaseCleaner with Capybara (https://github.com/DatabaseCleaner/database_cleaner#rspec-with-capybara-example) this document I believe describes the issue we face here (and possibly at other times).

We have the default config for DatabaseCleaner with rspec, but that assumes all connections to the database will be the same. If however you have anything which starts a new instance of the app and then tests against it, like using Capybara or invoking Rake, DatabaseCleaner does not have access to the connection.

This means anything created by the other instance is not cleaned. I also believe the same probably holds true for our `request` tests hence it looks like DatabaseCleaner is not always working.

To pin down the solution in this change we ran `be rspec --seed 34012` which would lead to

```bash
2) Reports::Boxi::RegistrationsSerializer#to_csv generates a csv file with data content
     Failure/Error: expect(result_lines.count).to eq(2)

       expected: 2
            got: 3

       (compared using ==)
     # ./spec/models/reports/boxi/registrations_serializer_spec.rb:21:in `block (3 levels) in <module:Boxi>'
     # ./spec/support/database_cleaner.rb:16:in `block (3 levels) in <top (required)>'
     # ./spec/support/database_cleaner.rb:15:in `block (2 levels) in <top (required)>'
```

With this change the count is back to being 3, though now we get a new error

```bash
2) RenewalLetterPresenter#listable_exemptions when there are 18 or fewer exemptions returns all exemptions
     Failure/Error: expect(subject.listable_exemptions).to eq(registration.exemptions)

       expected: #<ActiveRecord::Associations::CollectionProxy [#<WasteExemptionsEngine::Exemption id: 168, category: ...sing suitable spam rather ...", guidance: "This exemption allows you to use suitable spam rat...">]>
            got: [#<WasteExemptionsEngine::Exemption id: 166, category: 0, code: "F179", url: "https://example.gov.uk/...using suitable spam rather ...", guidance: "This exemption allows you to use suitable spam rat...">]

       (compared using ==)
```

This is why this is the first step. Next is to resolve the other failing test.