Add support for rake integration tests
This change will allow us to invoke rake tasks within out test suite to ensure that they run as expected.

Supporting rake integration tests comes with 2 primary challenges

- ensuring accurate test coverage
- ensuring any records created are correctly cleaned

Both result from the fact that invoking a rake task is the equivalent of starting a new instance. For simplecov this has the effect of it only capturing the coverage for the last rake task invoked. For DatabaseBase cleaner it is not called because the connection to the DB is new and not the one it is applying its transaction strategy to.

Each has its own solution. For simplecov we need to start an instance of Rake and load all tasks before the suite starts. This means its able to distinguish between which files are being called and collate the test coverage, rather than keep overwriting.

For DatabaseCleaner we need to mark our tests, and then use Rspec's config to alter the cleaning strategy when a rake test is called, plus clean the database after each rake spec.