Add support for rake integration tests (#429)
This change will allow us to invoke rake tasks within out test suite to ensure that they run as expected.

Supporting rake integration tests comes with 2 primary challenges

- ensuring accurate test coverage
- ensuring any records created are correctly cleaned

Both result from the fact that invoking a rake task is the equivalent of starting a new instance. For simplecov this has the effect of it only capturing the coverage for the last rake task invoked. For DatabaseBase cleaner it is not called because the connection to the DB is new and not the one it is applying its transaction strategy to.

Each has its own solution. For simplecov we need to start an instance of Rake and load all tasks before the suite starts. This means its able to distinguish between which files are being called and collate the test coverage, rather than keep overwriting.

For DatabaseCleaner we need to mark our tests, and then use Rspec's config to alter the cleaning strategy when a rake test is called, plus clean the database after each rake spec.

This involved a lot of head banging and working our way through to get things working, so below are some of the pertinent notes that were recorded when working on this change (in case its of use should things break!)

## Basic tests

Adding this support in this way causes Simplecov to start reporting no unit test coverage on the rake tasks which is a big hit to the test coverage. So we add basic tests for all existing rake tasks to nullify this issue.

## Skip an email test

We have chosen to skip the email test rake task. Essentially its not a test we actually want to run, so we mark it as skipped to have a permanent record. We worried that if we simply omitted the test we'd forgot and come back in the future thinking we needed to write one.

## Issue with the data

> This was the commit message for the change that resolved the rake tests leaving data that then broke other tests.

Now we are finding that the rake tests seem to be affecting others in cases where data is created. Any records created in the rake tests don't appear to be cleaned properly which means they remain and effect the results of others.

Though it talks about using DatabaseCleaner with Capybara (https://github.com/DatabaseCleaner/database_cleaner#rspec-with-capybara-example) this document I believe describes the issue we face here (and possibly at other times).

We have the default config for DatabaseCleaner with rspec, but that assumes all connections to the database will be the same. If however you have anything which starts a new instance of the app and then tests against it, like using Capybara or invoking Rake, DatabaseCleaner does not have access to the connection.

This means anything created by the other instance is not cleaned. I also believe the same probably holds true for our `request` tests hence it looks like DatabaseCleaner is not always working.

To pin down the solution in this change we ran `be rspec --seed 34012` which would lead to

```bash
2) Reports::Boxi::RegistrationsSerializer#to_csv generates a csv file with data content
     Failure/Error: expect(result_lines.count).to eq(2)

       expected: 2
            got: 3

       (compared using ==)
     # ./spec/models/reports/boxi/registrations_serializer_spec.rb:21:in `block (3 levels) in <module:Boxi>'
     # ./spec/support/database_cleaner.rb:16:in `block (3 levels) in <top (required)>'
     # ./spec/support/database_cleaner.rb:15:in `block (2 levels) in <top (required)>'
```

With this change the count is back to being 3, though now we get a new error

```bash
2) RenewalLetterPresenter#listable_exemptions when there are 18 or fewer exemptions returns all exemptions
     Failure/Error: expect(subject.listable_exemptions).to eq(registration.exemptions)

       expected: #<ActiveRecord::Associations::CollectionProxy [#<WasteExemptionsEngine::Exemption id: 168, category: ...sing suitable spam rather ...", guidance: "This exemption allows you to use suitable spam rat...">]>
            got: [#<WasteExemptionsEngine::Exemption id: 166, category: 0, code: "F179", url: "https://example.gov.uk/...using suitable spam rather ...", guidance: "This exemption allows you to use suitable spam rat...">]

       (compared using ==)
```