Update schedule.rb order by runtime (#430)
This is just some housekeeping of the `config/schedule.rb` class. We have ordered the schedule by runtime so its clearer to see in what order things run during the night for the WEX service.

We have also updated the comments to be consistent in their explanations, and how we set the time for each one.