Revert defra_ruby_validators to 2.1.2
We're not ready for [defra_ruby_validators](https://github.com/DEFRA/defra-ruby-validators) to be bumped yet, as we know it might break things in the app.

However we let Dependabot sneak it in, thinking it was just an engine update https://github.com/DEFRA/waste-exemptions-back-office/pull/435

So this change reverts it back to the previous version.