Update lookup tasks to use AssignSiteDetailsService
The service manages the calls to the other services in the engine that handle fetching the x & y and area for an address.

However it is intended only to assign the values so we need to add a call to `save!` and its good practise to use the bang version, which also means wrapping it in a being/rescue just in case.