Remove direct calls to CloseAirbrake
This is the pattern we wanted to follow in all Rake tasks. Should an error occur we wanted Airbrake to report it before the app instance closes. Now with defra-ruby-alert though, this will happen automatically everytime an app instance closes, whether it was instantiated via Rake or directly.

So this change removes all the existing references as they are no longer needed.