Switch airbrake mgmt to defra-ruby-alert (#500)
Whilst working on an issue for the [Flood Risk Activity Exemptions service](https://github.com/DEFRA/ruby-services-team/tree/master/services/frae) we came across an [issue](https://github.com/DEFRA/flood-risk-engine/pull/312) with Airbrake.

We wanted to add a new Rake task, and we wanted to apply our new pattern of ensuring Airbrake is closed when the task exits to ensure any errors are reported. The problem was due to the age of the Airbrake gem being used this caused an error in FRAE we hadn't seen in WEX and WCR.

So for reasons outlined in [FRAE PR 312](https://github.com/DEFRA/flood-risk-engine/pull/312) we came up with [defra-ruby-alert](https://github.com/DEFRA/defra-ruby-alert). The idea is going forward this will provide a single interface for configuring Airbrake and ensure the pattern of making sure all exceptions are reported before an app closes is applied consistently in all scenarios.

The main changes to enable this have been applied to the [waste-exemptions-engine](https://github.com/DEFRA/waste-exemptions-engine/pull/427).

These changes finalise the implementation for the back-office, switching out requiring and configuring Airbrake directly, to passing config details through to the WEX engine for it to do it.